project/backEnd/requirements.txt
# coding: utf-8

# cli 프로그램에서 오디오 재생을 위함
pygame==2.5.0

# 장고 웹프레임워크
# django==4.2.1

# 파이썬에서 보다 손쉬운 환경변수 처리
python-dotenv==1.0.0

# 장고 내에서 보다 손쉬운 환경변수 처리 (python-dotenv에 더해, 장고 친화적인 기능 지원)
django-environ==0.10.0

# openai api 사용을 위한 라이브러리
openai==0.27.7

# 장고 내에서 보다 손쉬운 bootstrap5 사용을 위한 라이브러리
django-bootstrap5==23.1

# HTTP 요청 및 HTML 파서 라이브러리
# requests
beautifulsoup4

# 장고 채널스 라이브러리
channels[daphne]~=4.0.0

# 구글번역 크롤링을 통한 음성 생성 라이브러리
gTTS==2.3.2


#######################

# DRF

asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.3
django-allauth==0.50.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
